+++ nepali/utils.py
@@ -14,12 +14,6 @@ def to_utc_timezone(datetime_obj):
 		datetime_obj = datetime_obj.replace(tzinfo=get_timezone())
 	return datetime_obj.astimezone(pytz.timezone('UTC'))
 
-def to_utc(datetime_obj):
-	warnings.warn(
-		message="to_utc is depreciated and no longer be available in version >= 1.0.0, use to_utc_timezone instead.", 
-		category=DeprecationWarning
-	)
-	return to_utc_timezone(datetime_obj)
 
 def to_nepali_timezone(datetime_obj):
 	if type(datetime_obj) != datetime.datetime:
@@ -29,12 +23,6 @@ def to_nepali_timezone(datetime_obj):
 		datetime_obj = datetime_obj.replace(tzinfo=get_timezone())
 	return datetime_obj.astimezone(NepaliTimeZone())
 
-def to_local(datetime_obj):
-	warnings.warn(
-		message="to_local is depreciated and no longer be available in version >= 1.0.0, use to_nepali_timezone instead.", 
-		category=DeprecationWarning
-	)
-	return to_nepali_timezone(datetime_obj)
 
 def to_nepalidatetime(datetime_object):
 	"""
@@ -54,12 +42,6 @@ def to_nepalidatetime(datetime_object):
 		return None
 	raise InvalidNepaliDateTimeObjectException('Argument must be instance of nepalidate or nepalidatetime or datetime.datetime or datetime.date') 
 
-def to_nepali_datetime(datetime_object):
-	warnings.warn(
-		message="to_nepali_datetime is depreciated and no longer be available in version >= 1.0.0, use to_nepalidatetime instead.", 
-		category=DeprecationWarning
-	)
-	return to_nepalidatetime(datetime_object)
 
 def to_nepalidate(datetime_object):
 	nepalidatetime_obj = to_nepalidatetime(datetime_object)